Fixed random numbers

+++ icc-x-api/icc-crypto-x-api.ts
@@ -29,7 +29,8 @@ export class IccCryptoXApi {
       c =>
         (
           Number(c) ^
-          (window.crypto.getRandomValues(new Uint8Array(1))![0] & (15 >> (Number(c) / 4)))
+          ((window.crypto.getRandomValues(new Uint8Array(1))! as Uint8Array)[0] &
+            (15 >> (Number(c) / 4)))
         ).toString(16) //Keep that inlined or you will loose the random
     )
   }
